+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Invalid input properly terminates Action execution (#1)
+- Compatibility with new `cargo-audit` JSON output (#70)
 
 ## [1.0.0] - 2019-10-09